# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Backend Base NestJS
  version: '1.0'
  description: Documentación Backend base - NestJS
  contact:
    name: agetic@agetic.gob.bo
servers:
  - url: 'http://localhost:3000'
  - url: 'https://dev.iop.agetic.gob.bo/backend-base'
  - url: 'https://base.agcs.agetic.gob.bo/ws'
paths:
...
  /api/autorizacion/politicas:
    get:
      summary: Api para obtener el listado de políticas
      tags: [ 'Autorización' ]
      operationId: get-api-autorizacion-politicas
      description: Api para obtener el listado de politicas
      parameters:
        - in: query
          name: tipo
          schema:
            type: string
            example: frontend
          description: tipo de política
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro(s) obtenido(s) con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      filas:
                        type: array
                        items:
                          type: object
                          properties:
                            sujeto:
                              type: string
                              example: 'ADMINISTRADOR'
                            objeto:
                              type: string
                              example: '/usuarios'
                            accion:
                              type: string
                              example: 'read'
                            app:
                              type: string
                              example: 'frontend'
                      total:
                        type: number
                        example: 1
    post:
      summary: Api para crear una nueva política
      tags: [ 'Autorización' ]
      operationId: post-api-autorizacion-politicas
      description: Api para crear una nueva política
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sujeto:
                  type: string
                  example: 'ADMINISTRADOR'
                objeto:
                  type: string
                  example: 'recurso'
                accion:
                  type: string
                  example: 'read'
                app:
                  type: string
                  example: 'frontend'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro creado con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      sujeto:
                        type: string
                        example: 'ADMINISTRADOR'
                      objeto:
                        type: string
                        example: 'recurso'
                      accion:
                        type: string
                        example: 'read'
                      app:
                        type: string
                        example: 'frontend'
    delete:
      summary: Api para eliminar una política
      tags: [ 'Autorización' ]
      operationId: delete-api-autorizacion-politicas
      description: Api para eliminar una política
      responses:
        '204':
          description: NOT_CONTENT
      parameters:
        - in: query
          name: sujeto
          schema:
            type: string
            example: ADMINISTRADOR
          description: Nombre del rol (SUJETO)
        - in: query
          name: objeto
          schema:
            type: string
            example: RECURSO
          description: Nombre del recurso (OBJETO)
        - in: query
          name: accion
          schema:
            type: string
            example: GET
          description: Nombre de la acción (ACCIÓN)
        - in: query
          name: app
          schema:
            type: string
            example: Backend
          description: Nombre de la aplicación
  /api/parametros:
    get:
      summary: Api para obtener el listado de parámetros
      tags: [ 'Parámetros' ]
      operationId: get-api-parametros
      description: Api para obtener el listado de parámetros
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro(s) obtenido(s) con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      filas:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                            codigo:
                              type: string
                              example: 'TD-1'
                            nombre:
                              type: string
                              example: 'Tipo de documento CI'
                            grupo:
                              type: string
                              example: 'TD'
                      total:
                        type: number
                        example: 1
    post:
      summary: Api para crear un nuevo parámetro
      tags: [ 'Parámetros' ]
      operationId: post-api-parametros
      description: Api para crear un nuevo parámetro
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                codigo:
                  type: string
                  example: 'TD-2'
                nombre:
                  type: string
                  example: 'Documento de extranjería'
                grupo:
                  type: string
                  example: 'TD'
                descripcion:
                  type: string
                  example: 'Tipo de documento de extranjería'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro creado con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      codigo:
                        type: string
                        example: 'TD-1'
                      nombre:
                        type: string
                        example: 'Documento extranjería'
                      grupo:
                        type: string
                        example: 'TD'
                      descripcion:
                        type: string
                        example: 'Tipo de documento de extranjería'
                      id:
                        type: string
                        example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                      estado:
                        type: string
                        example: 'ACTIVO'
  /api/parametros/{grupo}/listado:
    get:
      summary: Api para obtener el listado de parámetros por grupo
      tags: [ 'Parámetros' ]
      operationId: get-api-parametros-grupo
      description: Api para obtener el listado de parámetros por grupo
      parameters:
        - in: path
          name: grupo
          required: true
          schema:
            type: string
            example: 'TD'
          description: grupo de parámetros
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro(s) obtenido(s) con éxito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                        nombre:
                          type: string
                          example: 'Cédula de identidad'
                        codigo:
                          type: string
                          example: 'TD-1'
...